Add package init files and cleanup addon layout
