+++ README.md
@@ -22,7 +22,8 @@ blidge/
 ├── panels/               # UI パネル
 ├── renderer/             # ビューポート描画関連
 ├── ui/                   # UIList など
-└── utils/                # 補助スクリプト
+├── utils/                # 補助スクリプト
+└── lib/                  # 依存ライブラリを配置
 ```
 
 各モジュールの主な役割は次の通りです。
@@ -32,6 +33,7 @@ blidge/
 - **panels** : 3D View や Properties に表示される UI パネルを提供します。
 - **renderer** : 仮想メッシュを描画するためのシェーダーやジオメトリデータを持ちます。
 - **utils** : シーン解析、glTF 設定取得、WebSocket サーバー等の補助機能です。
+各ディレクトリには `__init__.py` を追加し、アドオンを ZIP 化してもパッケージとして認識されるよう整理しました。
 
 ## 依存ライブラリ
